Fixing warnings in IE driver for 64-bit compilation

+++ cpp/iedriver/CommandHandlers/ExecuteAsyncScriptCommandHandler.cpp
@@ -58,7 +58,7 @@ void ExecuteAsyncScriptCommandHandler::ExecuteInternal(
 
     Json::Value json_args = args_parameter_iterator->second;
 
-    int timeout_value = executor.async_script_timeout();
+    unsigned long long timeout_value = executor.async_script_timeout();
     std::wstring timeout = std::to_wstring(static_cast<long long>(timeout_value));
 
     std::wstring script_body = StringUtilities::ToWString(script_parameter_iterator->second.asString());
